// src/types/token-result.ts
export type TokenResult = {
    google_access_token?: string | undefined;
    google_expires_in?: string | undefined;
    google_refresh_token?: string | undefined;
    google_state: string;
    google_token_error?: string | undefined;
    google_token_error_description?: string | undefined;
    google_token_error_uri?: string | undefined;
    google_error?: string | undefined;
    google_error_description?: string | undefined;
    google_error_uri?: string | undefined;
}